added stats and update some changes in the post update

+++ src/modules/Post/post.service.ts
@@ -1,4 +1,4 @@
-import { CommentStatus, Post } from "../../generated/prisma/client";
+import { CommentStatus, Post, PostStatus } from "../../generated/prisma/client";
 import { PostWhereInput } from "../../generated/prisma/models";
 import { prisma } from "../../lib/prisma";
 
@@ -104,40 +104,65 @@ const getPosts = async ({
   };
 };
 
+const getOwnPost = async (userId: string) => {
+  await prisma.user.findFirstOrThrow({
+    where: {
+      id: userId,
+      status: "ACTIVE",
+    },
+    select: {
+      id: true,
+    },
+  });
+
+  const result = await prisma.post.findMany({
+    where: {
+      authorId: userId,
+    },
+    orderBy: {
+      createdAt: "desc",
+    },
+  });
+
+  return result;
+};
+
 // get single post
-const getSinglePost = async (id: string) => {
+const getPostById = async (postId: string) => {
   return await prisma.$transaction(async (tx) => {
+    const postExists = await tx.post.findUnique({
+      where: { id: postId },
+      select: { id: true },
+    });
+
+    if (!postExists) {
+      throw new Error("Post not found");
+    }
+
     await tx.post.update({
-      where: {
-        id,
-      },
+      where: { id: postId },
       data: {
         views: {
           increment: 1,
         },
       },
     });
+
     const postData = await tx.post.findUnique({
-      where: { id },
+      where: { id: postId },
       include: {
         comments: {
           where: {
             parentId: null,
             status: CommentStatus.APPROVED,
           },
-          orderBy: {
-            createdAt: "desc",
-          },
+          orderBy: { createdAt: "desc" },
           include: {
             replies: {
-              where: {
-                status: CommentStatus.APPROVED,
-              },
+              where: { status: CommentStatus.APPROVED },
               include: {
                 replies: {
-                  where: {
-                    status: CommentStatus.APPROVED,
-                  },
+                  where: { status: CommentStatus.APPROVED },
                 },
               },
             },
@@ -150,12 +175,35 @@ const getSinglePost = async (id: string) => {
         },
       },
     });
+
     return postData;
   });
 };
 
 // update post
-const updatePost = async (id: string, payload: any) => {
+const updatePost = async (
+  id: string,
+  payload: any,
+  authorId: string,
+  isAdmin: boolean
+) => {
+  const postData = await prisma.post.findFirstOrThrow({
+    where: {
+      id,
+    },
+    select: {
+      id: true,
+      authorId: true,
+    },
+  });
+  if (postData.authorId !== authorId) {
+    throw new Error("This post is not yours");
+  }
+
+  if (!isAdmin) {
+    delete payload.isFeatured;
+  }
+
   const result = await prisma.post.update({
     where: { id },
     data: payload,
@@ -164,17 +212,80 @@ const updatePost = async (id: string, payload: any) => {
 };
 
 // delete post
-const deletePost = async (id: string) => {
+const deletePost = async (
+  postId: string,
+  authorId: string,
+  isAdmin: boolean
+) => {
+  const postData = await prisma.post.findFirstOrThrow({
+    where: {
+      id: postId,
+    },
+    select: {
+      id: true,
+      authorId: true,
+    },
+  });
+
+  if (!isAdmin && postData.authorId !== authorId) {
+    throw new Error("This post is not yours");
+  }
+
   const result = await prisma.post.delete({
-    where: { id },
+    where: { id: postId },
   });
   return result;
+};
+
+// stats
+const stats = async () => {
+  return await prisma.$transaction(async (tx) => {
+    const [
+      totalPost,
+      publishedPost,
+      archivedPost,
+      draftPost,
+      totalComment,
+      approvedComment,
+      totalUser,
+      totalAdmin,
+      userCount,
+      totalViewsInPost,
+    ] = await Promise.all([
+      await tx.post.count(),
+      await tx.post.count({ where: { status: PostStatus.PUBLISHED } }),
+      await tx.post.count({ where: { status: PostStatus.ARCHIVED } }),
+      await tx.post.count({ where: { status: PostStatus.DRAFT } }),
+      await tx.comment.count(),
+      await tx.comment.count({ where: { status: CommentStatus.APPROVED } }),
+      await tx.user.count(),
+      await tx.user.count({ where: { role: "ADMIN" } }),
+      await tx.user.count({ where: { role: "USER" } }),
+      await tx.post.aggregate({
+        _sum: { views: true },
+      }),
+    ]);
+    return {
+      totalPost,
+      publishedPost,
+      archivedPost,
+      draftPost,
+      totalComment,
+      approvedComment,
+      totalUser,
+      totalAdmin,
+      userCount,
+      totalViewsInPost: totalViewsInPost._sum.views,
+    };
+  });
 };
 
 export const postService = {
   createPost,
   getPosts,
-  getSinglePost,
+  getPostById,
+  getOwnPost,
   updatePost,
   deletePost,
+  stats,
 };
